Change how the x-axis for degree-Rips bifiltrations is parameterized, and change the x-axis label as well.
I need this for a theoretical paper I am writing.  I think this parameterization is probably the better one in general, particularly when comparing random samples of different sizes.  Perhaps we should consider using this in the main RIVET code, but I don't have a strong opinion about this.

TODO: My solution for changing the x-labels, in data_reader.cpp, is a bit hacky; though that hack was already present in the code.  I used the same hack a second time.